Final version(s) of my CLT program from EAS4240
Final and Final_Q4 refer to the final exam of the course, and thus the final versions of the code written. I didn't get a perfect score, although I was close, so there are some issues. I am not completely certain where or what they are but I have reason to believe they are either related to methods I used to answer the final exam questions specifically or in the calculation of concepts covered closest to the end of the course, since I didn't have as much time to refine those.